Added user path and links to it to view user's recent posts
